requirements.txt
biopython
regex
tqdm
pandas